Refactor: consolidate utils and setup
