+++ src/vulkan_app/mod.rs
@@ -4,8 +4,6 @@ pub const HEIGHT: u32 = 600;
 pub use app::VulkanApp;
 
 mod app;
-mod debug;
-mod queue;
-mod swapchain_support;
-mod ubo;
+mod utils;
+mod setup;
 mod vertex;
