+++ src/vulkan_app/setup.rs
@@ -1,0 +1,282 @@
+use ash::{vk, Entry};
+use winit::window::Window;
+use std::ffi::{CStr, CString};
+
+use super::utils::{QueueFamilyIndices, SwapchainSupportDetails, UniformBufferObject, vulkan_debug_callback};
+use super::vertex::Vertex;
+
+pub(super) fn create_instance(entry: &Entry, window: &Window) -> ash::Instance {
+    let app_name = CString::new("Vulkan Triangle").unwrap();
+    let engine_name = CString::new("No Engine").unwrap();
+    let app_info = vk::ApplicationInfo::builder()
+        .application_name(&app_name)
+        .application_version(vk::make_api_version(0, 1, 0, 0))
+        .engine_name(&engine_name)
+        .engine_version(vk::make_api_version(0, 1, 0, 0))
+        .api_version(vk::API_VERSION_1_0);
+
+    let mut extension_names =
+        ash_window::enumerate_required_extensions(window.raw_display_handle())
+            .unwrap()
+            .to_vec();
+    extension_names.push(ash::extensions::ext::DebugUtils::name().as_ptr());
+
+    let create_info = vk::InstanceCreateInfo::builder()
+        .application_info(&app_info)
+        .enabled_extension_names(&extension_names);
+
+    unsafe {
+        entry
+            .create_instance(&create_info, None)
+            .expect("Failed to create instance")
+    }
+}
+
+pub(super) fn setup_debug_messenger(
+    entry: &Entry,
+    instance: &ash::Instance,
+) -> (ash::extensions::ext::DebugUtils, vk::DebugUtilsMessengerEXT) {
+    let debug_info = vk::DebugUtilsMessengerCreateInfoEXT::builder()
+        .message_severity(
+            vk::DebugUtilsMessageSeverityFlagsEXT::ERROR
+                | vk::DebugUtilsMessageSeverityFlagsEXT::WARNING,
+        )
+        .message_type(
+            vk::DebugUtilsMessageTypeFlagsEXT::GENERAL
+                | vk::DebugUtilsMessageTypeFlagsEXT::VALIDATION
+                | vk::DebugUtilsMessageTypeFlagsEXT::PERFORMANCE,
+        )
+        .pfn_user_callback(Some(vulkan_debug_callback));
+
+    let debug_utils_loader = ash::extensions::ext::DebugUtils::new(entry, instance);
+    let debug_messenger = unsafe {
+        debug_utils_loader
+            .create_debug_utils_messenger(&debug_info, None)
+            .unwrap()
+    };
+
+    (debug_utils_loader, debug_messenger)
+}
+
+pub(super) fn create_render_pass(
+    device: &ash::Device,
+    format: vk::Format,
+    depth_format: vk::Format,
+) -> vk::RenderPass {
+    let color_attachment = vk::AttachmentDescription::builder()
+        .format(format)
+        .samples(vk::SampleCountFlags::TYPE_1)
+        .load_op(vk::AttachmentLoadOp::CLEAR)
+        .store_op(vk::AttachmentStoreOp::STORE)
+        .stencil_load_op(vk::AttachmentLoadOp::DONT_CARE)
+        .stencil_store_op(vk::AttachmentStoreOp::DONT_CARE)
+        .initial_layout(vk::ImageLayout::UNDEFINED)
+        .final_layout(vk::ImageLayout::PRESENT_SRC_KHR);
+
+    let color_attachment_ref = vk::AttachmentReference::builder()
+        .attachment(0)
+        .layout(vk::ImageLayout::COLOR_ATTACHMENT_OPTIMAL);
+
+    let depth_attachment = vk::AttachmentDescription::builder()
+        .format(depth_format)
+        .samples(vk::SampleCountFlags::TYPE_1)
+        .load_op(vk::AttachmentLoadOp::CLEAR)
+        .store_op(vk::AttachmentStoreOp::DONT_CARE)
+        .stencil_load_op(vk::AttachmentLoadOp::DONT_CARE)
+        .stencil_store_op(vk::AttachmentStoreOp::DONT_CARE)
+        .initial_layout(vk::ImageLayout::UNDEFINED)
+        .final_layout(vk::ImageLayout::DEPTH_STENCIL_ATTACHMENT_OPTIMAL);
+
+    let depth_attachment_ref = vk::AttachmentReference::builder()
+        .attachment(1)
+        .layout(vk::ImageLayout::DEPTH_STENCIL_ATTACHMENT_OPTIMAL);
+
+    let subpass = vk::SubpassDescription::builder()
+        .pipeline_bind_point(vk::PipelineBindPoint::GRAPHICS)
+        .color_attachments(std::slice::from_ref(&color_attachment_ref))
+        .depth_stencil_attachment(&depth_attachment_ref);
+
+    let dependency = vk::SubpassDependency::builder()
+        .src_subpass(vk::SUBPASS_EXTERNAL)
+        .dst_subpass(0)
+        .src_stage_mask(
+            vk::PipelineStageFlags::COLOR_ATTACHMENT_OUTPUT
+                | vk::PipelineStageFlags::EARLY_FRAGMENT_TESTS,
+        )
+        .src_access_mask(vk::AccessFlags::empty())
+        .dst_stage_mask(
+            vk::PipelineStageFlags::COLOR_ATTACHMENT_OUTPUT
+                | vk::PipelineStageFlags::EARLY_FRAGMENT_TESTS,
+        )
+        .dst_access_mask(
+            vk::AccessFlags::COLOR_ATTACHMENT_WRITE
+                | vk::AccessFlags::DEPTH_STENCIL_ATTACHMENT_WRITE,
+        );
+
+    let attachments = [color_attachment.build(), depth_attachment.build()];
+    let render_pass_info = vk::RenderPassCreateInfo::builder()
+        .attachments(&attachments)
+        .subpasses(std::slice::from_ref(&subpass))
+        .dependencies(std::slice::from_ref(&dependency));
+
+    unsafe { device.create_render_pass(&render_pass_info, None).unwrap() }
+}
+
+pub(super) fn create_buffer(
+    instance: &ash::Instance,
+    device: &ash::Device,
+    pdevice: vk::PhysicalDevice,
+    size: vk::DeviceSize,
+    usage: vk::BufferUsageFlags,
+    properties: vk::MemoryPropertyFlags,
+) -> (vk::Buffer, vk::DeviceMemory) {
+    let buffer_info = vk::BufferCreateInfo::builder()
+        .size(size)
+        .usage(usage)
+        .sharing_mode(vk::SharingMode::EXCLUSIVE);
+
+    let buffer = unsafe { device.create_buffer(&buffer_info, None).unwrap() };
+    let mem_requirements = unsafe { device.get_buffer_memory_requirements(buffer) };
+    let mem_type_index = find_memory_type(instance, pdevice, mem_requirements.memory_type_bits, properties);
+
+    let alloc_info = vk::MemoryAllocateInfo::builder()
+        .allocation_size(mem_requirements.size)
+        .memory_type_index(mem_type_index);
+
+    let buffer_memory = unsafe { device.allocate_memory(&alloc_info, None).unwrap() };
+    unsafe {
+        device.bind_buffer_memory(buffer, buffer_memory, 0).unwrap();
+    }
+
+    (buffer, buffer_memory)
+}
+
+pub(super) fn find_memory_type(
+    instance: &ash::Instance,
+    pdevice: vk::PhysicalDevice,
+    type_filter: u32,
+    properties: vk::MemoryPropertyFlags,
+) -> u32 {
+    let mem_properties = unsafe { instance.get_physical_device_memory_properties(pdevice) };
+    for i in 0..mem_properties.memory_type_count {
+        if (type_filter & (1 << i)) != 0
+            && mem_properties.memory_types[i as usize].property_flags.contains(properties)
+        {
+            return i;
+        }
+    }
+    panic!("Failed to find suitable memory type!");
+}
+
+pub(super) fn create_image(
+    instance: &ash::Instance,
+    device: &ash::Device,
+    pdevice: vk::PhysicalDevice,
+    width: u32,
+    height: u32,
+    format: vk::Format,
+    tiling: vk::ImageTiling,
+    usage: vk::ImageUsageFlags,
+    properties: vk::MemoryPropertyFlags,
+) -> (vk::Image, vk::DeviceMemory) {
+    let image_info = vk::ImageCreateInfo::builder()
+        .image_type(vk::ImageType::TYPE_2D)
+        .extent(vk::Extent3D { width, height, depth: 1 })
+        .mip_levels(1)
+        .array_layers(1)
+        .format(format)
+        .tiling(tiling)
+        .initial_layout(vk::ImageLayout::UNDEFINED)
+        .usage(usage)
+        .sharing_mode(vk::SharingMode::EXCLUSIVE)
+        .samples(vk::SampleCountFlags::TYPE_1);
+
+    let image = unsafe { device.create_image(&image_info, None).unwrap() };
+
+    let mem_requirements = unsafe { device.get_image_memory_requirements(image) };
+    let mem_type_index = find_memory_type(instance, pdevice, mem_requirements.memory_type_bits, properties);
+
+    let alloc_info = vk::MemoryAllocateInfo::builder()
+        .allocation_size(mem_requirements.size)
+        .memory_type_index(mem_type_index);
+
+    let image_memory = unsafe { device.allocate_memory(&alloc_info, None).unwrap() };
+    unsafe {
+        device.bind_image_memory(image, image_memory, 0).unwrap();
+    }
+
+    (image, image_memory)
+}
+
+pub(super) fn create_image_view(
+    device: &ash::Device,
+    image: vk::Image,
+    format: vk::Format,
+    aspect_flags: vk::ImageAspectFlags,
+) -> vk::ImageView {
+    let view_info = vk::ImageViewCreateInfo::builder()
+        .image(image)
+        .view_type(vk::ImageViewType::TYPE_2D)
+        .format(format)
+        .subresource_range(vk::ImageSubresourceRange {
+            aspect_mask: aspect_flags,
+            base_mip_level: 0,
+            level_count: 1,
+            base_array_layer: 0,
+            layer_count: 1,
+        });
+
+    unsafe { device.create_image_view(&view_info, None).unwrap() }
+}
+
+pub(super) fn find_supported_format(
+    instance: &ash::Instance,
+    pdevice: vk::PhysicalDevice,
+    candidates: &[vk::Format],
+    tiling: vk::ImageTiling,
+    features: vk::FormatFeatureFlags,
+) -> vk::Format {
+    for &format in candidates {
+        let props = unsafe { instance.get_physical_device_format_properties(pdevice, format) };
+
+        if tiling == vk::ImageTiling::LINEAR && props.linear_tiling_features.contains(features) {
+            return format;
+        } else if tiling == vk::ImageTiling::OPTIMAL && props.optimal_tiling_features.contains(features) {
+            return format;
+        }
+    }
+    panic!("Failed to find supported format!");
+}
+
+pub(super) fn find_depth_format(instance: &ash::Instance, pdevice: vk::PhysicalDevice) -> vk::Format {
+    find_supported_format(
+        instance,
+        pdevice,
+        &[vk::Format::D32_SFLOAT, vk::Format::D32_SFLOAT_S8_UINT, vk::Format::D24_UNORM_S8_UINT],
+        vk::ImageTiling::OPTIMAL,
+        vk::FormatFeatureFlags::DEPTH_STENCIL_ATTACHMENT,
+    )
+}
+
+pub(super) fn create_depth_resources(
+    instance: &ash::Instance,
+    device: &ash::Device,
+    pdevice: vk::PhysicalDevice,
+    extent: vk::Extent2D,
+) -> (vk::Image, vk::DeviceMemory, vk::ImageView) {
+    let depth_format = find_depth_format(instance, pdevice);
+    let (depth_image, depth_image_memory) = create_image(
+        instance,
+        device,
+        pdevice,
+        extent.width,
+        extent.height,
+        depth_format,
+        vk::ImageTiling::OPTIMAL,
+        vk::ImageUsageFlags::DEPTH_STENCIL_ATTACHMENT,
+        vk::MemoryPropertyFlags::DEVICE_LOCAL,
+    );
+    let depth_image_view = create_image_view(device, depth_image, depth_format, vk::ImageAspectFlags::DEPTH);
+
+    (depth_image, depth_image_memory, depth_image_view)
+}
