--- src/vulkan_app/swapchain_support.rs
@@ -1,7 +1,0 @@
-use ash::vk;
-
-pub struct SwapchainSupportDetails {
-    pub capabilities: vk::SurfaceCapabilitiesKHR,
-    pub formats: Vec<vk::SurfaceFormatKHR>,
-    pub present_modes: Vec<vk::PresentModeKHR>,
-}
