+++ src/vulkan_app/utils.rs
@@ -1,0 +1,47 @@
+use ash::vk;
+use cgmath::Matrix4;
+use std::ffi::CStr;
+
+#[derive(Clone, Copy)]
+pub struct QueueFamilyIndices {
+    pub graphics_family: Option<u32>,
+    pub present_family: Option<u32>,
+}
+
+impl QueueFamilyIndices {
+    pub fn new() -> Self {
+        Self {
+            graphics_family: None,
+            present_family: None,
+        }
+    }
+
+    pub fn is_complete(&self) -> bool {
+        self.graphics_family.is_some() && self.present_family.is_some()
+    }
+}
+
+pub struct SwapchainSupportDetails {
+    pub capabilities: vk::SurfaceCapabilitiesKHR,
+    pub formats: Vec<vk::SurfaceFormatKHR>,
+    pub present_modes: Vec<vk::PresentModeKHR>,
+}
+
+#[derive(Copy, Clone)]
+pub struct UniformBufferObject {
+    pub model: Matrix4<f32>,
+    pub view: Matrix4<f32>,
+    pub proj: Matrix4<f32>,
+}
+
+pub unsafe extern "system" fn vulkan_debug_callback(
+    message_severity: vk::DebugUtilsMessageSeverityFlagsEXT,
+    _message_type: vk::DebugUtilsMessageTypeFlagsEXT,
+    p_callback_data: *const vk::DebugUtilsMessengerCallbackDataEXT,
+    _p_user_data: *mut std::ffi::c_void,
+) -> vk::Bool32 {
+    let callback_data = *p_callback_data;
+    let message = CStr::from_ptr(callback_data.p_message).to_string_lossy();
+    println!("{:?}: {}", message_severity, message);
+    vk::FALSE
+}
